[react] feat: Improved error handling when pattern drafting fails (#205)
If drafting a part fails, the logs store contains an array [message, stacktrace] instead of a simple string. Handle this in the LogView.

Also display an error message in the DraftView when there were errors during drafting.

![image](/attachments/62cdebe3-7825-4382-88fd-395337cda58b)

after clicking the button:

![image](/attachments/f909887b-b9bd-467e-9e13-f850c38d76b3)

(could also display the stack traces, but they're quite unhelpful and you can find them in the browser console anyway, the code for that is commented out)

Closes #163

Reviewed-on: https://codeberg.org/freesewing/freesewing/pulls/205
Reviewed-by: Joost De Cock <joostdecock@noreply.codeberg.org>
Co-authored-by: Jonathan Haas <haasjona@gmail.com>
Co-committed-by: Jonathan Haas <haasjona@gmail.com>