fix(react): Untranslated strings also need a trailing space (#349)
The bug appears in part title cutlist instructions. The bug behavior exists only in the React component. It does not exist with the core SVG renderer.

Before:
![Screenshot 2025-05-12 at 10.59.49 AM.png](/attachments/cb43d749-bad3-4d54-95ba-ae1ed4675bd2)

After:
![Screenshot 2025-05-12 at 10.59.15 AM.png](/attachments/979ca23a-1db4-49e6-8b31-4e2cd84ecbdf)

Edit: (Added, in case the keyword needs to be in the PR description/first comment, not just in a subsequent comment):
Fixes #331

Co-authored-by: Benjamin Fan <ben-git@swinglonga.com>
Reviewed-on: https://codeberg.org/freesewing/freesewing/pulls/349
Reviewed-by: Joost De Cock <joostdecock@noreply.codeberg.org>
Co-authored-by: Benjamin Fan <benjamesben@noreply.codeberg.org>
Co-committed-by: Benjamin Fan <benjamesben@noreply.codeberg.org>