fix(react): Pattern clone fix and improvment (#353)
This PR:
1. Fixes an issue where the pattern clone operation fails, resulting in a hang/freeze.
2. Appends " (clone)" text to the cloned pattern's name.

The hang/freeze problem was reported on Discord:
https://discord.com/channels/698854858052075530/944926520282054666/1372209603080753212
discord://discord.com/channels/698854858052075530/944926520282054666/1372209603080753212

Because a successful pattern clone operation immediately loads the new, cloned pattern in the browser, the loading status "success" message is not seen by the user. Instead the user sees the new pattern, but because the new pattern looks exactly like the old pattern, it is not immediately clear to the user that the operation succeeded.

This PR adds " (clone)" to the cloned pattern's name, making it easier to see that the clone operation succeeded and than the user is actually looking at the new, cloned pattern. It also helps differentiate the new, cloned pattern from the original pattern.

Co-authored-by: Benjamin Fan <ben-git@swinglonga.com>
Reviewed-on: https://codeberg.org/freesewing/freesewing/pulls/353
Reviewed-by: Joost De Cock <joostdecock@noreply.codeberg.org>
Co-authored-by: Benjamin Fan <benjamesben@noreply.codeberg.org>
Co-committed-by: Benjamin Fan <benjamesben@noreply.codeberg.org>