[dev] fix: Amended the example usage for cbqc (#445)
The code currently used in the note about `cqbc` only approximates a quarter circle when `radius==1`. As far as I understand, `cqbc` is a ratio, and ought to be multiplied by the intended radius of curvature. Doing this appears to produce good results at any scale, so I'm suggesting a change to the example.

I've also added the relevant `import` to make things slightly easier. (As beginners to JS modules, we ended up cribbing from Onyx to work this out!)

Reviewed-on: https://codeberg.org/freesewing/freesewing/pulls/445
Reviewed-by: Joost De Cock <joostdecock@noreply.codeberg.org>
Co-authored-by: Ben Morris <bugs@benmorris.org.uk>
Co-committed-by: Ben Morris <bugs@benmorris.org.uk>