feat: Better support for snapped percentage options
The adapts the Sabrina design to switch a mm option to a snapped
percentage option.

In addition to that, it adds the new `snappedPctOption` utility
method to core as a named export. This method can wrap the option
config and takes a measurement name to achieve the same result as
`pctBasedOn` but for snapped percentage options.

Finally, it adapts the Editor component to add a note that the
absolute value is snapped.